+++ plugins/qcheck-stm/src/ortac_qcheck_stm.ml
@@ -4,14 +4,14 @@ module Ir_of_gospel = Ir_of_gospel
 module Reserr = Reserr
 module Stm_of_ir = Stm_of_ir
 
-let main path init sut output quiet () =
+let main path init sut include_ output quiet () =
   let open Reserr in
   let fmt = Registration.get_out_formatter output in
   let pp = pp quiet Ppxlib_ast.Pprintast.structure fmt in
   pp
     (let* sigs, config = Config.init path init sut in
      let* ir = Ir_of_gospel.run sigs config in
-     Stm_of_ir.stm config ir)
+     Stm_of_ir.stm include_ config ir)
 
 open Cmdliner
 
@@ -38,10 +38,24 @@ end = struct
              system under test."
           ~docv:"INIT")
 
+  let include_ =
+    Arg.(
+      value
+      & opt (some string) None
+      & info [ "i"; "include" ] ~docv:"MODULE"
+          ~doc:"Include MODULE in the generated code.")
+
   let term =
     let open Registration in
     Term.(
-      const main $ ocaml_file $ init $ sut $ output_file $ quiet $ setup_log)
+      const main
+      $ ocaml_file
+      $ init
+      $ sut
+      $ include_
+      $ output_file
+      $ quiet
+      $ setup_log)
 
   let cmd = Cmd.v info term
 end